moy-11 relocated transactions
