+++ projects/moyney/src/app/home/home.component.html
@@ -5,6 +5,6 @@
 
   <moy-card [config]="cards.recently_added">
     <moy-table [config]="recentlyAdded"></moy-table>
-    <aside *ngIf="!recentlyAdded.matrix.length">No data</aside>
+    <aside *ngIf="!recentlyAdded.matrix.length" class="Home__empty-list">Hey! you haven't added anything yet!</aside>
   </moy-card>
 </section>
