+++ projects/moyney/src/app/home/home.component.scss
@@ -7,3 +7,9 @@
     color: lightcoral;
   }
 }
+
+.Home__empty-list {
+  font-size: 0.8rem;
+  padding: 0.5rem;
+  text-align: center;
+}
