+++ projects/moyney/src/app/home/home.module.ts
@@ -2,9 +2,9 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { MoyCardModule } from '@libs/moy-card/moy-card.module';
 import { MoyTableModule } from '@libs/moy-table/moy-table.module';
+import { TransactionModule } from '../transaction/transaction.module';
 import { HomeRoutingModule } from './home-routing.module';
 import { HomeComponent } from './home.component';
-import { TransactionModule } from './transaction/transaction.module';
 
 @NgModule({
   imports: [CommonModule, HomeRoutingModule, TransactionModule, MoyCardModule, MoyTableModule],
